docs: add DashedLineChart example in TypeScript (.tsx)
